refactor(core): Move `typeorm` operators from `CredentialsService` to repositories (no-changelog)
